Rename component folder
